A drawing function is added to the sketch.
When people get closer than a 1.5 m to te device, a beat is played to make them engage with the device. The volume gets louder the closer u get to the device.
If you press a button, the note is played and a splash is drawn at a random location with a colour specific to the instrument.